Inject props to remote component
