+++ ui/components/NavigatorExtension.js
@@ -14,9 +14,10 @@ import LoadingScreen from "./LoadingComponents/LoadingComponent";
 import usePreventUserFromLeavingPage from "../utils/hooks/usePreventUserFromLeavingPage";
 import { getK8sClusterIdsFromCtxId } from "../utils/multi-ctx";
 import ConfirmationModal from "./ConfirmationModal"
-import { getComponentsinFile } from "../utils/utils";
+import { getComponentsinFile, generateValidatePayload } from "../utils/utils";
 import UploadImport from "./UploadImport";
 import ConfigurationSubscription from "../components/graphql/subscriptions/ConfigurationSubscription";
+import Validation from "./Validation";
 
 const requires = createRequires(getDependencies);
 const useRemoteComponent = createUseRemoteComponent({ requires });
@@ -66,7 +67,9 @@ function Extension({ grafana, prometheus, updateLoadTestData, url, isDrawerColla
         },
         ConfirmationModal,
         getComponentsinFile,
-        UploadImport
+        UploadImport,
+        generateValidatePayload,
+        Validation
       }}
     />
   );
